requirements.txt
PyYAML>=6.0.1
pytest>=7.0.0

